# network-path: memory and CPU optimization to `parseTCP`