<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?

In some profiles we can see that the network path traceroute feature is, in some cases, allocating a lot of memory: [example profile](https://ddstaging.datadoghq.com/profiling/explorer?query=service%3Asystem-probe%20kube_cluster_name%3Astingchameleon%20&agg_m=count&agg_m_source=base&agg_t=count&color_by=package&cols=timestamp%2Cservice%2C%40metrics.core_cpu_cores%2C%40metrics.core_alloc_bytes_per_sec%2Cversion%2Chost%2C%40metrics.go_lifetime_heap_bytes&event=AgAAAZKWXUtjprEfJQAAAAAAAAAYAAAAAEFaS1dYVXV4QUFCdmtHZWY0Q2Q3REFBQQAAACQAAAAAMDE5Mjk2NWUtYzk0NC00OTk3LWEwYWMtYjE1YWRhYTI5MGRl&fromUser=true&group_by=line&limit=100&my_code=disabled&profile_type=alloc-size&profileId=AZKWXUuxAABvkGef4Cd7DAAA&stream_sort=%40metrics.core_alloc_bytes_per_sec%2Cdesc&viz=stream&start=1729097200736&end=1729100800736&paused=true).

The goal of this PR is to improve the performance of the `parseTCP` function from an allocation standpoint but also from a CPU usage point of view.

Benchmarks:
```
$ go test -benchmem -run=^$ -tags test -bench ^BenchmarkParseTCP$ github.com/DataDog/datadog-agent/pkg/networkpath/traceroute/tcp
# before
BenchmarkParseTCP-10    	 3401749	       338.8 ns/op	     816 B/op	      13 allocs/op
# after
BenchmarkParseTCP-10    	10998318	       103.4 ns/op	     384 B/op	       1 allocs/op

# benchstat
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/networkpath/traceroute/tcp
cpu: Apple M1 Max
            │   old.txt    │               new.txt               │
            │    sec/op    │   sec/op     vs base                │
ParseTCP-10   334.90n ± 1%   95.70n ± 2%  -71.43% (p=0.000 n=10)

            │  old.txt   │              new.txt               │
            │    B/op    │    B/op     vs base                │
ParseTCP-10   816.0 ± 0%   384.0 ± 0%  -52.94% (p=0.000 n=10)

            │   old.txt   │              new.txt               │
            │  allocs/op  │ allocs/op   vs base                │
ParseTCP-10   13.000 ± 0%   1.000 ± 0%  -92.31% (p=0.000 n=10)
```

The main things this PR does, is move the things that are not moving, that are re-usable from one call to `parseTCP` to the next to a structure, owned by `handlePackets`. This allows `parseTCP` to reduce drastically the amount of allocation required.

### Motivation

### Describe how to test/QA your changes

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->